Movie_App/requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
psycopg2==2.9.1
djangorestframework-simplejwt==4.7.2
django-cors-headers==3.7.0
Pillow==8.3.1
coverage==5.5
drf-writable-nested
celery==5.1
django-celery-results==2.2.0
requests